pyproject.toml
[project]
name = "sync-buddy"
version = "2022.0a1"
dependencies = [
    "HiYaPyCo",
    "SQLAlchemy",
    "schema",
    "requests",
    "python-dotenv",
    "jsonpath-ng",
    "rauth"
]

[project.optional-dependencies]
dev = [
    "autopep8",
    "pytest",
    "pytest-cov",
    "pytest-mock"
]
